# docs/phase5/logs/openapi-v1-phase5.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "DerasaX.Api",
    "version": "1.0"
  },
  "paths": {
...
    "/api/v1/ai/analysis": {
      "post": {
        "tags": [
          "AiAnalysis"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AiDto.GenerateAnalysisDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AiDto.GenerateAnalysisDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AiDto.GenerateAnalysisDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ai/analysis/{studentId}/history": {
      "get": {
        "tags": [
          "AiAnalysis"
        ],
        "parameters": [
          {
            "name": "studentId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/v1/ai/analysis/{painPointId}/review": {
      "put": {
        "tags": [
          "AiAnalysis"
        ],
        "parameters": [
          {
            "name": "painPointId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AiDto.ReviewPainPointDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AiDto.ReviewPainPointDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/DerasaX.Application.Dto.AiDto.ReviewPainPointDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DerasaX.Application.Dto.AiDto.AnalysisTurnDto": {
        "type": "object",
        "properties": {
          "role": {
            "type": "string",
            "nullable": true
          },
          "content": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "DerasaX.Application.Dto.AiDto.GenerateAnalysisDto": {
        "type": "object",
        "properties": {
          "studentId": {
            "type": "string",
            "nullable": true
          },
          "conversation": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DerasaX.Application.Dto.AiDto.AnalysisTurnDto"
            },
            "nullable": true
          },
          "subject": {
            "type": "string",
            "nullable": true
          },
          "unitId": {
            "type": "string",
            "nullable": true
          },
          "lessonId": {
            "type": "string",
            "nullable": true
          },
          "language": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "DerasaX.Application.Dto.AiDto.ReviewPainPointDto": {
        "type": "object",
        "properties": {
          "decision": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "Bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}